fix: wrap useSearchParams usage in Suspense to fix Next.js static rendering error
