+++ frontend/src/app/oauth-success/page.js
@@ -1,50 +1,12 @@
-"use client";
+import { Suspense } from "react";
+import OAuthSuccess from "./OAuthSuccess";
 
-import { useEffect } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import { useError } from "@/contexts/ErrorContext";
+export const dynamic = "force-dynamic"; 
 
-const OAuthSuccess = () => {
-  const searchParams = useSearchParams();
-  const router = useRouter();
-  const { showError } = useError();
-  
-  useEffect(() => {
-    // Get URL parameters
-    const error = searchParams.get('error');
-    const message = searchParams.get('message');
-    const token = searchParams.get('token');
-    
-    // Handle error case
-    if (error && message) {
-      // Redirect to home with error parameters
-      router.push(`/?error=${error}&message=${message}`);
-      return;
-    }
-    
-    // Handle success case
-    if (token) {
-      // Just redirect to home page as we're using httpOnly cookies
-      router.push('/');
-    } else {
-      // No token found, redirect to signin
-      router.push('/signin');
-    }
-  }, [searchParams, router, showError]);
-
+export default function OAuthSuccessPage() {
   return (
-    <div className="flex justify-center items-center min-h-screen bg-[#edead7]">
-      <div className="flex flex-col items-center space-y-6">
-        <div className="relative w-14 h-14 opacity-100">
-          <div className="absolute inset-0 rounded-full border-t-4 border-r-4 border-[#d56c4e] animate-spin"/>
-          <div className="absolute inset-1 rounded-full bg-[#edead7]"/>
-        </div>
-        <p className="text-xl font-medium tracking-wide text-gray-800">
-          Processing your login...
-        </p>
-      </div>
-    </div>
+    <Suspense fallback={<div>Loading...</div>}>
+      <OAuthSuccess />
+    </Suspense>
   );
-};
-
-export default OAuthSuccess;
+}
